pyproject.toml
[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"

[project]
name = "robot_framework"
version = "0.0.1"
authors = [
  { name="ITK Development", email="itk-rpa@mkb.aarhus.dk" },
]
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
]
dependencies = [
    "itk-dev-shared-components == 2.*",
    "OpenOrchestrator == 1.*",
    "Pillow == 9.*",
    "selenium == 4.*",
    "requests == 2.*",
    "openpyxl == 3.*",
    "python-dotenv"
]

[project.optional-dependencies]
dev = [
  "pylint",
  "flake8"
]